Fix NPE in EmptyTopDocsCollectorContext::createManager
